+++ src/main/java/inter/domain/InterTransaction.java
@@ -11,10 +11,10 @@ public class InterTransaction implements Serializable {
     private String fromAccount;
     private String toBankCode;
     private String fromBankCode;
-    private long amount;
+    private double amount;
     private String fromReference;
 
-    public InterTransaction(String toAccount, String fromAccount, long amount, String fromReference) {
+    public InterTransaction(String toAccount, String fromAccount, double amount, String fromReference) {
         AccountCodeUtil accountCodeUtil = new AccountCodeUtil();
 
         this.toAccount = toAccount;
@@ -42,7 +42,7 @@ public class InterTransaction implements Serializable {
         return fromBankCode;
     }
 
-    public long getAmount() {
+    public double getAmount() {
         return amount;
     }
